fix: cache is being stored in the incorrect location
The user_cache_dir() function returns the directory without the trailing
slash so the path where this is being stored is incorrect because its
not correctly building up the path. Switched to Python's pathlib library
and used path building they've got. Switched all calls to using it as it
has more guardrails than os.path does.

Signed-off-by: Doug Goldstein <cardoe@cardoe.com>